Cleanup: common base error type, doc fixes
Previously, almost all errors returned were just errors.New and started
"spillover:", which is not a match for the chosen package name.  The
GitHub repo has no watchers and no stargazers, so I'm breaking compat
here.

Now, all errors generated within this package are either bubbled through
unmodified or are of type `SwallowError`, and the string of those is now
explicitly guaranteed to start `swallowjson:`.

Removed a stray debugging print to stdout in an error case which should
not have slipped through.  Can't quite remove the `fmt` dependency,
because it's the sanest way to repr a `json.Token`.

Removed two TODO items.

Improved documentation a little.